FINERACT-1932: Fineract modularization - fineract-branch
