+++ build.gradle
@@ -30,6 +30,7 @@ buildscript {
                 'fineract-core',
                 'fineract-accounting',
                 'fineract-provider',
+                'fineract-branch',
                 'fineract-investor',
                 'fineract-loan',
                 'fineract-savings',
@@ -52,7 +53,8 @@ buildscript {
                 'fineract-provider',
                 'fineract-investor',
                 'fineract-loan',
-                'fineract-savings'
+                'fineract-savings',
+                'fineract-branch'
             ].contains(it.name)
         }
         npmRepository = 'https://npm.pkg.github.com'
