+++ fineract-branch/src/main/java/org/apache/fineract/organisation/teller/data/CashierTransactionTypeTotalsData.java
