+++ fineract-branch/src/main/java/org/apache/fineract/organisation/teller/handler/CreateTellerCommandHandler.java
@@ -18,41 +18,30 @@
  */
 package org.apache.fineract.organisation.teller.handler;
 
+import lombok.RequiredArgsConstructor;
 import org.apache.fineract.commands.annotation.CommandType;
 import org.apache.fineract.commands.handler.NewCommandSourceHandler;
 import org.apache.fineract.infrastructure.core.api.JsonCommand;
 import org.apache.fineract.infrastructure.core.data.CommandProcessingResult;
 import org.apache.fineract.organisation.teller.service.TellerWritePlatformService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 /**
- * Handles a delete cashier command.
+ * Handles a create teller command.
  *
  * @author Markus Geiss
- * @see org.apache.fineract.organisation.teller.service.CashierWritePlatformService
+ * @see org.apache.fineract.organisation.teller.service.TellerWritePlatformService
  * @since 2.0.0
  */
 @Service
-@CommandType(entity = "TELLER", action = "DELETECASHIERALLOCATION")
-public class DeleteCashierAllocationCommandHandler implements NewCommandSourceHandler {
+@RequiredArgsConstructor
+@CommandType(entity = "TELLER", action = "CREATE")
+public class CreateTellerCommandHandler implements NewCommandSourceHandler {
 
     private final TellerWritePlatformService writePlatformService;
 
-    /**
-     * Creates a new instance.
-     *
-     * @param writePlatformService
-     *            the {@code CashierWritePlatformService} used to access the backend
-     */
-    @Autowired
-    public DeleteCashierAllocationCommandHandler(final TellerWritePlatformService writePlatformService) {
-
-        this.writePlatformService = writePlatformService;
-    }
-
     @Override
     public CommandProcessingResult processCommand(final JsonCommand command) {
-        return this.writePlatformService.deleteCashierAllocation(command.entityId(), command.subentityId(), command);
+        return this.writePlatformService.createTeller(command);
     }
 }
