+++ fineract-branch/src/main/java/org/apache/fineract/organisation/teller/handler/CreateTellerTransactionCommandHandler.java
@@ -18,35 +18,19 @@
  */
 package org.apache.fineract.organisation.teller.handler;
 
-import org.apache.fineract.commands.annotation.CommandType;
+import lombok.RequiredArgsConstructor;
 import org.apache.fineract.commands.handler.NewCommandSourceHandler;
 import org.apache.fineract.infrastructure.core.api.JsonCommand;
 import org.apache.fineract.infrastructure.core.data.CommandProcessingResult;
-import org.apache.fineract.organisation.teller.service.TellerWritePlatformService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+import org.apache.fineract.organisation.teller.service.TellerTransactionWritePlatformService;
 
-@Service
-@CommandType(entity = "TELLER", action = "SETTLECASHFROMCASHIER")
-public class SettleCashFromCashierCommandHandler implements NewCommandSourceHandler {
+@RequiredArgsConstructor
+public class CreateTellerTransactionCommandHandler implements NewCommandSourceHandler {
 
-    private final TellerWritePlatformService writePlatformService;
-
-    /**
-     * Creates a new instance.
-     *
-     * @param writePlatformService
-     *            the {@code CashierWritePlatformService} used to access the backend
-     */
-    @Autowired
-    public SettleCashFromCashierCommandHandler(final TellerWritePlatformService writePlatformService) {
-
-        this.writePlatformService = writePlatformService;
-    }
+    private final TellerTransactionWritePlatformService writePlatformService;
 
     @Override
     public CommandProcessingResult processCommand(final JsonCommand command) {
-
-        return this.writePlatformService.settleCashFromCashier(command.subentityId(), command);
+        return this.writePlatformService.createTellerTransaction(command);
     }
 }
