+++ fineract-branch/src/main/java/org/apache/fineract/organisation/teller/handler/ModifyCashierCommandHandler.java
@@ -18,35 +18,26 @@
  */
 package org.apache.fineract.organisation.teller.handler;
 
-import org.apache.fineract.commands.annotation.CommandType;
+import lombok.RequiredArgsConstructor;
 import org.apache.fineract.commands.handler.NewCommandSourceHandler;
 import org.apache.fineract.infrastructure.core.api.JsonCommand;
 import org.apache.fineract.infrastructure.core.data.CommandProcessingResult;
-import org.apache.fineract.organisation.teller.service.TellerWritePlatformService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+import org.apache.fineract.organisation.teller.service.CashierWritePlatformService;
 
-@Service
-@CommandType(entity = "TELLER", action = "ALLOCATECASHIER")
-public class AllocateCashierToTellerCommandHandler implements NewCommandSourceHandler {
+/**
+ * Handles a modify cashier command.
+ *
+ * @author Markus Geiss
+ * @see org.apache.fineract.organisation.teller.service.CashierWritePlatformService
+ * @since 2.0.0
+ */
+@RequiredArgsConstructor
+public class ModifyCashierCommandHandler implements NewCommandSourceHandler {
 
-    private final TellerWritePlatformService writePlatformService;
-
-    /**
-     * Creates a new instance.
-     *
-     * @param writePlatformService
-     *            the {@code CashierWritePlatformService} used to access the backend
-     */
-    @Autowired
-    public AllocateCashierToTellerCommandHandler(final TellerWritePlatformService writePlatformService) {
-
-        this.writePlatformService = writePlatformService;
-    }
+    private final CashierWritePlatformService writePlatformService;
 
     @Override
     public CommandProcessingResult processCommand(final JsonCommand command) {
-
-        return this.writePlatformService.allocateCashierToTeller(command.entityId(), command);
+        return this.writePlatformService.modifyCashier(command.entityId(), command);
     }
 }
