+++ fineract-branch/src/main/java/org/apache/fineract/organisation/teller/handler/SettleCashFromCashierCommandHandler.java
@@ -18,40 +18,24 @@
  */
 package org.apache.fineract.organisation.teller.handler;
 
+import lombok.RequiredArgsConstructor;
 import org.apache.fineract.commands.annotation.CommandType;
 import org.apache.fineract.commands.handler.NewCommandSourceHandler;
 import org.apache.fineract.infrastructure.core.api.JsonCommand;
 import org.apache.fineract.infrastructure.core.data.CommandProcessingResult;
 import org.apache.fineract.organisation.teller.service.TellerWritePlatformService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-/**
- * Handles a create teller command.
- *
- * @author Markus Geiss
- * @see org.apache.fineract.organisation.teller.service.TellerWritePlatformService
- * @since 2.0.0
- */
 @Service
-@CommandType(entity = "TELLER", action = "CREATE")
-public class CreateTellerCommandHandler implements NewCommandSourceHandler {
+@RequiredArgsConstructor
+@CommandType(entity = "TELLER", action = "SETTLECASHFROMCASHIER")
+public class SettleCashFromCashierCommandHandler implements NewCommandSourceHandler {
 
     private final TellerWritePlatformService writePlatformService;
 
-    /**
-     * Creates a new instance
-     *
-     * @param writePlatformService
-     *            the {@code TellerWritePlatformService} used to access the backend
-     */
-    @Autowired
-    public CreateTellerCommandHandler(final TellerWritePlatformService writePlatformService) {
-        this.writePlatformService = writePlatformService;
-    }
-
     @Override
     public CommandProcessingResult processCommand(final JsonCommand command) {
-        return this.writePlatformService.createTeller(command);
+
+        return this.writePlatformService.settleCashFromCashier(command.subentityId(), command);
     }
 }
