+++ fineract-provider/dependencies.gradle
@@ -29,6 +29,7 @@ dependencies {
     implementation(project(path: ':fineract-investor'))
     implementation(project(path: ':fineract-loan'))
     implementation(project(path: ':fineract-savings'))
+    implementation(project(path: ':fineract-branch'))
     providedRuntime("org.springframework.boot:spring-boot-starter-tomcat")
 
     // Never use "compile" scope, but make all dependencies either 'implementation', 'runtimeOnly' or 'testCompile'.
