+++ settings.gradle
@@ -51,6 +51,7 @@ rootProject.name='fineract'
 include ':fineract-core'
 include ':fineract-accounting'
 include ':fineract-provider'
+include ':fineract-branch'
 include ':fineract-investor'
 include ':fineract-loan'
 include ':fineract-savings'
